feat(cart): 장바구니 조회 서비스 로직 구현 및 금액 타입 Long으로 변경
- CartService getCart 로직 추가
- Redis 장바구니 데이터를 조회해 상품 정보와 매핑하도록 구현
- 상품별 수량, 금액, 총 금액 계산 로직 추가
- 장바구니 key 형식을 cart:member:{memberId}로 변경
- Cart/Order 금액 필드와 Service의 계산 로직을 Long 기반으로 변경

+++ src/main/java/one/theone/server/domain/cart/dto/response/CartItemResponse.java
@@ -5,8 +5,8 @@ import java.math.BigDecimal;
 public record CartItemResponse (
         Long productId,
         String productName,
-        BigDecimal price,
+        Long price,
         Integer quantity,
-        BigDecimal lineAmount
+        Long lineAmount
 ){
 }
